# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: Starknet Keyring Proxy API
  version: 0.1.0
components:
  securitySchemes:
    hmacHeaders:
      type: apiKey
      in: header
      name: X-Keyring-Signature
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /v1/sign/session-transaction:
    post:
      summary: Sign a Starknet session transaction payload
      description: |
        Requires HMAC headers:
        - X-Keyring-Timestamp
        - X-Keyring-Nonce
        - X-Keyring-Signature
      security:
        - hmacHeaders: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - accountAddress
                - chainId
                - nonce
                - validUntil
                - calls
              properties:
                accountAddress:
                  type: string
                keyId:
                  type: string
                  description: Optional signer key id. If omitted, server default key is used.
                chainId:
                  type: string
                nonce:
                  type: string
                validUntil:
                  type: integer
                calls:
                  type: array
                  items:
                    type: object
                    required: [contractAddress, entrypoint, calldata]
                    properties:
                      contractAddress:
                        type: string
                      entrypoint:
                        type: string
                      calldata:
                        type: array
                        items:
                          type: string
      responses:
        '200':
          description: Signature produced
        '401':
          description: Authentication failed
        '422':
          description: Policy violation (self-call, denied selector, invalid chainId, or invalid validUntil)